backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.16,<2.0
python-jose[cryptography]>=3.5,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt<5
pydantic-settings>=2.10,<3.0
python-multipart>=0.0.20,<1.0
pytest>=8.4,<9.0
